Update background.js
Add some handling for redirects that dont need to occur.
Before using the redirect value, we now test the original URL with a /robots.txt endpoint. If this returns 200 we go with that. Seems to help retrieve some sites that are not fully disabled but have been redirected to new pages.